+++ CHANGELOG.md
@@ -17,6 +17,10 @@ Emojis for the following are chosen based on [gitmoji](https://gitmoji.dev/).
 - Checks for `aria-label` and `alt-text` terminal punctuation was expanded to handle more languages ([#108](https://github.com/activist-org/i18n-check/issues/108)).
 - The total number of failed checks is reported to the user.
 
+### 🐛 Bug Fixes
+
+- The `key-formatting` and `key-naming` `--fix` options now also update key usage in configured `nonexistent-keys` `search-dirs` ([#118](https://github.com/activist-org/i18n-check/issues/118)).
+
 ### ♻️ Code Refactoring
 
 - The legacy Python typing system was removed in favor of the more modern approach ([#113](https://github.com/activist-org/i18n-check/issues/113)).